requirements.txt
altgraph==0.17.3
contourpy==1.1.0
cycler==0.11.0
EasyProcess==1.1
entrypoint2==1.1
evdev==1.6.1
fonttools==4.42.1
jeepney==0.8.0
kiwisolver==1.4.5
MouseInfo==0.1.3
mss==9.0.1
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytesseract==0.3.10
python-dateutil==2.8.2
python-xlib==0.33
python3-xlib==0.15
pytweening==1.0.7
six==1.16.0
